ask before install c++ and programs

+++ Windows-Settings.ps1
@@ -896,68 +896,101 @@ Write-Host "Done."
 ############################################################################################################################################################
 <# C++ Installation #>
 ############################################################################################################################################################
-Write-Host "Installing C++ for running applications" 
+$install = Read-Host "Do you want to install C++ Redistributable dependencies for games? (Y/N)"
+if ($install -match "[Yy]") {
+    Write-Host "Installing C++ for running applications"
 
     function Get-FileFromWeb {
-    param ([Parameter(Mandatory)][string]$URL, [Parameter(Mandatory)][string]$File)
-    function Show-Progress {
-    param ([Parameter(Mandatory)][Single]$TotalValue, [Parameter(Mandatory)][Single]$CurrentValue, [Parameter(Mandatory)][string]$ProgressText, [Parameter()][int]$BarSize = 10, [Parameter()][switch]$Complete)
-    $percent = $CurrentValue / $TotalValue
-    $percentComplete = $percent * 100
-    if ($psISE) { Write-Progress "$ProgressText" -id 0 -percentComplete $percentComplete }
-    else { Write-Host -NoNewLine "`r$ProgressText $(''.PadRight($BarSize * $percent, [char]9608).PadRight($BarSize, [char]9617)) $($percentComplete.ToString('##0.00').PadLeft(6)) % " }
+        param (
+            [Parameter(Mandatory)][string]$URL, 
+            [Parameter(Mandatory)][string]$File
+        )
+        function Show-Progress {
+            param (
+                [Parameter(Mandatory)][Single]$TotalValue, 
+                [Parameter(Mandatory)][Single]$CurrentValue, 
+                [Parameter(Mandatory)][string]$ProgressText, 
+                [Parameter()][int]$BarSize = 10, 
+                [Parameter()][switch]$Complete
+            )
+            $percent = $CurrentValue / $TotalValue
+            $percentComplete = $percent * 100
+            if ($psISE) { 
+                Write-Progress "$ProgressText" -id 0 -percentComplete $percentComplete 
+            } else { 
+                Write-Host -NoNewLine "`r$ProgressText $(''.PadRight($BarSize * $percent, [char]9608).PadRight($BarSize, [char]9617)) $($percentComplete.ToString('##0.00').PadLeft(6)) % " 
+            }
+        }
+        try {
+            $request = [System.Net.HttpWebRequest]::Create($URL)
+            $response = $request.GetResponse()
+            if ($response.StatusCode -eq 401 -or $response.StatusCode -eq 403 -or $response.StatusCode -eq 404) { 
+                throw "Remote file either doesn't exist, is unauthorized, or is forbidden for '$URL'." 
+            }
+            if ($File -match '^\.\\') { 
+                $File = Join-Path (Get-Location -PSProvider 'FileSystem') ($File -Split '^\.')[1] 
+            }
+            if ($File -and !(Split-Path $File)) { 
+                $File = Join-Path (Get-Location -PSProvider 'FileSystem') $File 
+            }
+            if ($File) { 
+                $fileDirectory = $([System.IO.Path]::GetDirectoryName($File)); 
+                if (!(Test-Path($fileDirectory))) { 
+                    [System.IO.Directory]::CreateDirectory($fileDirectory) | Out-Null 
+                } 
+            }
+            [long]$fullSize = $response.ContentLength
+            [byte[]]$buffer = new-object byte[] 1048576
+            [long]$total = [long]$count = 0
+            $reader = $response.GetResponseStream()
+            $writer = new-object System.IO.FileStream $File, 'Create'
+            do {
+                $count = $reader.Read($buffer, 0, $buffer.Length)
+                $writer.Write($buffer, 0, $count)
+                $total += $count
+                if ($fullSize -gt 0) { Show-Progress -TotalValue $fullSize -CurrentValue $total -ProgressText " $($File.Name)" }
+            } while ($count -gt 0)
+        }
+        finally {
+            $reader.Close()
+            $writer.Close()
+        }
     }
-    try {
-    $request = [System.Net.HttpWebRequest]::Create($URL)
-    $response = $request.GetResponse()
-    if ($response.StatusCode -eq 401 -or $response.StatusCode -eq 403 -or $response.StatusCode -eq 404) { throw "Remote file either doesn't exist, is unauthorized, or is forbidden for '$URL'." }
-    if ($File -match '^\.\\') { $File = Join-Path (Get-Location -PSProvider 'FileSystem') ($File -Split '^\.')[1] }
-    if ($File -and !(Split-Path $File)) { $File = Join-Path (Get-Location -PSProvider 'FileSystem') $File }
-    if ($File) { $fileDirectory = $([System.IO.Path]::GetDirectoryName($File)); if (!(Test-Path($fileDirectory))) { [System.IO.Directory]::CreateDirectory($fileDirectory) | Out-Null } }
-    [long]$fullSize = $response.ContentLength
-    [byte[]]$buffer = new-object byte[] 1048576
-    [long]$total = [long]$count = 0
-    $reader = $response.GetResponseStream()
-    $writer = new-object System.IO.FileStream $File, 'Create'
-    do {
-    $count = $reader.Read($buffer, 0, $buffer.Length)
-    $writer.Write($buffer, 0, $count)
-    $total += $count
-    if ($fullSize -gt 0) { Show-Progress -TotalValue $fullSize -CurrentValue $total -ProgressText " $($File.Name)" }
-    } while ($count -gt 0)
+
+    # Download and install C++ Redistributables
+    $files = @{
+        "https://download.microsoft.com/download/8/B/4/8B42259F-5D70-43F4-AC2E-4B208FD8D66A/vcredist_x86.EXE" = "$env:TEMP\vcredist2005_x86.exe"
+        "https://download.microsoft.com/download/8/B/4/8B42259F-5D70-43F4-AC2E-4B208FD8D66A/vcredist_x64.EXE" = "$env:TEMP\vcredist2005_x64.exe"
+        "https://download.microsoft.com/download/5/D/8/5D8C65CB-C849-4025-8E95-C3966CAFD8AE/vcredist_x86.exe" = "$env:TEMP\vcredist2008_x86.exe"
+        "https://download.microsoft.com/download/5/D/8/5D8C65CB-C849-4025-8E95-C3966CAFD8AE/vcredist_x64.exe" = "$env:TEMP\vcredist2008_x64.exe"
+        "https://download.microsoft.com/download/1/6/5/165255E7-1014-4D0A-B094-B6A430A6BFFC/vcredist_x86.exe" = "$env:TEMP\vcredist2010_x86.exe"
+        "https://download.microsoft.com/download/1/6/5/165255E7-1014-4D0A-B094-B6A430A6BFFC/vcredist_x64.exe" = "$env:TEMP\vcredist2010_x64.exe"
+        "https://download.microsoft.com/download/1/6/B/16B06F60-3B20-4FF2-B699-5E9B7962F9AE/VSU_4/vcredist_x86.exe" = "$env:TEMP\vcredist2012_x86.exe"
+        "https://download.microsoft.com/download/1/6/B/16B06F60-3B20-4FF2-B699-5E9B7962F9AE/VSU_4/vcredist_x64.exe" = "$env:TEMP\vcredist2012_x64.exe"
+        "https://aka.ms/highdpimfc2013x86enu" = "$env:TEMP\vcredist2013_x86.exe"
+        "https://aka.ms/highdpimfc2013x64enu" = "$env:TEMP\vcredist2013_x64.exe"
+        "https://aka.ms/vs/17/release/vc_redist.x86.exe" = "$env:TEMP\vcredist2015_2017_2019_2022_x86.exe"
+        "https://aka.ms/vs/17/release/vc_redist.x64.exe" = "$env:TEMP\vcredist2015_2017_2019_2022_x64.exe"
     }
-    finally {
-    $reader.Close()
-    $writer.Close()
+
+    foreach ($url in $files.Keys) {
+        Get-FileFromWeb -URL $url -File $files[$url]
     }
+
+    # Start installers
+    $arguments = @("/q", "/qb", "/passive /norestart")
+    $files.Values | ForEach-Object {
+        $argIndex = 0
+        if ($_ -match "2008") { $argIndex = 1 }
+        elseif ($_ -match "2010|2012|2013|2015|2017|2019|2022") { $argIndex = 2 }
+        Start-Process -wait $_ -ArgumentList $arguments[$argIndex]
     }
-# download c++ installers
-Get-FileFromWeb -URL "https://download.microsoft.com/download/8/B/4/8B42259F-5D70-43F4-AC2E-4B208FD8D66A/vcredist_x86.EXE" -File "$env:TEMP\vcredist2005_x86.exe"
-Get-FileFromWeb -URL "https://download.microsoft.com/download/8/B/4/8B42259F-5D70-43F4-AC2E-4B208FD8D66A/vcredist_x64.EXE" -File "$env:TEMP\vcredist2005_x64.exe"
-Get-FileFromWeb -URL "https://download.microsoft.com/download/5/D/8/5D8C65CB-C849-4025-8E95-C3966CAFD8AE/vcredist_x86.exe" -File "$env:TEMP\vcredist2008_x86.exe"
-Get-FileFromWeb -URL "https://download.microsoft.com/download/5/D/8/5D8C65CB-C849-4025-8E95-C3966CAFD8AE/vcredist_x64.exe" -File "$env:TEMP\vcredist2008_x64.exe"
-Get-FileFromWeb -URL "https://download.microsoft.com/download/1/6/5/165255E7-1014-4D0A-B094-B6A430A6BFFC/vcredist_x86.exe" -File "$env:TEMP\vcredist2010_x86.exe" 
-Get-FileFromWeb -URL "https://download.microsoft.com/download/1/6/5/165255E7-1014-4D0A-B094-B6A430A6BFFC/vcredist_x64.exe" -File "$env:TEMP\vcredist2010_x64.exe"
-Get-FileFromWeb -URL "https://download.microsoft.com/download/1/6/B/16B06F60-3B20-4FF2-B699-5E9B7962F9AE/VSU_4/vcredist_x86.exe" -File "$env:TEMP\vcredist2012_x86.exe"
-Get-FileFromWeb -URL "https://download.microsoft.com/download/1/6/B/16B06F60-3B20-4FF2-B699-5E9B7962F9AE/VSU_4/vcredist_x64.exe" -File "$env:TEMP\vcredist2012_x64.exe"
-Get-FileFromWeb -URL "https://aka.ms/highdpimfc2013x86enu" -File "$env:TEMP\vcredist2013_x86.exe"
-Get-FileFromWeb -URL "https://aka.ms/highdpimfc2013x64enu" -File "$env:TEMP\vcredist2013_x64.exe"
-Get-FileFromWeb -URL "https://aka.ms/vs/17/release/vc_redist.x86.exe" -File "$env:TEMP\vcredist2015_2017_2019_2022_x86.exe"
-Get-FileFromWeb -URL "https://aka.ms/vs/17/release/vc_redist.x64.exe" -File "$env:TEMP\vcredist2015_2017_2019_2022_x64.exe"
-# start c++ installers
-Start-Process -wait "$env:TEMP\vcredist2005_x86.exe" -ArgumentList "/q"
-Start-Process -wait "$env:TEMP\vcredist2005_x64.exe" -ArgumentList "/q"
-Start-Process -wait "$env:TEMP\vcredist2008_x86.exe" -ArgumentList "/qb"
-Start-Process -wait "$env:TEMP\vcredist2008_x64.exe" -ArgumentList "/qb"
-Start-Process -wait "$env:TEMP\vcredist2010_x86.exe" -ArgumentList "/passive /norestart"
-Start-Process -wait "$env:TEMP\vcredist2010_x64.exe" -ArgumentList "/passive /norestart"
-Start-Process -wait "$env:TEMP\vcredist2012_x86.exe" -ArgumentList "/passive /norestart"
-Start-Process -wait "$env:TEMP\vcredist2012_x64.exe" -ArgumentList "/passive /norestart"
-Start-Process -wait "$env:TEMP\vcredist2013_x86.exe" -ArgumentList "/passive /norestart"
-Start-Process -wait "$env:TEMP\vcredist2013_x64.exe" -ArgumentList "/passive /norestart"
-Start-Process -wait "$env:TEMP\vcredist2015_2017_2019_2022_x86.exe" -ArgumentList "/passive /norestart"
-Start-Process -wait "$env:TEMP\vcredist2015_2017_2019_2022_x64.exe" -ArgumentList "/passive /norestart"
-Write-Host "Done." 
+
+    Write-Host "Done."
+} else {
+    Write-Host "Installation canceled."
+}
+
 ############################################################################################################################################################
 <# Direct X Installation #>
 ############################################################################################################################################################
@@ -1049,17 +1082,27 @@ Get-WindowsPackage -Online | Where PackageName -like *QuickAssist* | Remove-Wind
 ############################################################################################################################################################
 # Personal Programs
 ############################################################################################################################################################
-Write-Output "Installing Chocolatey"
-choco install chocolatey-core.extension -y
-Write-Output "Installing Personal Applications"
-choco install discord.install -y
-choco install notepadplusplus.install -y
-choco install peazip.install -y
-choco install vlc.install -y
-choco install steam -y
-choco install ddu -y
-choco install tightvnc -y
-choco install rustdesk -y
+# Prompt the user
+$installPersonalApps = Read-Host "Do you want to install personal applications before running? (Y/N)"
+
+# If the user responds with Y or y, proceed with the installation of personal applications
+if ($installPersonalApps -eq 'Y' -or $installPersonalApps -eq 'y') {
+    Write-Output "Installing Chocolatey"
+    choco install chocolatey-core.extension -y
+
+    Write-Output "Installing Personal Applications"
+    choco install discord.install -y
+    choco install notepadplusplus.install -y
+    choco install peazip.install -y
+    choco install vlc.install -y
+    choco install steam -y
+    choco install ddu -y
+    choco install tightvnc -y
+    choco install rustdesk -y
+} else {
+    Write-Output "Skipping Personal Applications installation."
+}
+
 #Write-Output "Running Spotify Install Script"
 #iex "& { $(iwr -useb 'https://raw.githubusercontent.com/SpotX-Official/spotx-official.github.io/main/run.ps1') } -new_theme"
